wpebackend: new package
This is a base library needed by the WPE WebKit port. It contains a
loader for different "backends", which in turn implement platform
specific access to hardware. This mechanism is used mainly by the
graphics subsystem (hence the need for EGL headers) and input devices
support.

Based on patches originally written by:

- Zan Dobersek <zdobersek@igalia.com>
- albertd <albert@dahan.nl>
- wouterlucas <wouter@wouterlucas.com>

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>