wpewebkit: new package
This is a recipe to build the WebKit WPE port, based off the WebKitGTK+
one with a few changes. Additional options allowing further configuration
of features enabled at build-time to reduce used disk and memory usage
may be added later on.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>

---

Note that the location and version numbering for the tarball is
temporary. The WPE team (me included) is working on setting up an
official releases channel, so I would send a patch later on to fix
that, or re-upload a new version of this one.

Also, regarding the list of supported architectures, Debian does have
builds for more platforms like PowerPC and S390 and I am confident that
there are people using the packages because there is the occasional bug
report in the Debian bug tracker. If anybody here has access to the
relevant hardware and can build and test, I will be happy to provide
assistance and upstreaming patches to WebKit both for WebKitGTK+ and
WPE. You can check the list of Debian architectures here:

  https://packages.debian.org/sid/libwebkit2gtk-4.0-37