wpebackend-rdk: new package
WPE backend developed as part of the RDK platform. It supports a few
different build modes. This only enables building with support for the
Raspberry Pi VideoCore graphics chip (using dispmanx), or for using the
generic Wayland-EGL interface. The dependencies needed for building this
WPE backend for the rest of hardwre it supports are not available at the
moment in Buildroot.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>