wpewebkit: Remove the multimedia package
The multimedia package is valuable, because multimedia is not needed on
some systems, and pulls in a huge amount of dependencies. We should
definitely keep it.

But it's also totally broken due to a recursive dependency issue, and
I'm not sure how to fix it. There is no circular dependency... it seems
BuildStream is just not really great about ordering dependencies.

Without some fix for this issue, it is simply not possible to build the
wpewebkit module. So, remove the multimedia package for now. We should
find a way to restore it that does not prevent the main module from
being built.